killed MinGW warnings (many thanks to Lei HY!)

+++ include/cinolib/gui/qt/surface_mesh_control_panel.cpp
@@ -665,7 +665,7 @@ void SurfaceMeshControlPanel<Mesh>::connect()
         {
             ScalarField sf(filename.c_str());
             sf.normalize_in_01();
-            if (sf.size() == m->num_verts()) sf.copy_to_mesh(*m);
+            if ((uint)sf.size() == m->num_verts()) sf.copy_to_mesh(*m);
             else std::cerr << "Could not load scalar field " << filename << " - array size mismatch!" << endl;
         }
     });
